Allow QtVariantContext to be used with either QVariantHash or QVariantMap
QHash is generally preferable to QMap for unordered dictionaries as it
is faster, however there are a number of places in Qt and external
libraries which use QVariantMap (such as QJson).

Use QVariantHash in the mustache tests and examples, but allow a
QtVariantContext to work with either a QVariantHash or a QVariantMap.