Add tests from Mustache specification v1.1.2.
This adds a new test TestMustache::testConformance(). The test is
driven by JSON data from version 1.1.2 of the Mustache spec [1],
which has been copied to the tests/specs directory. The test is
only built when Qt 5.0.0+ is available, since it uses the JSON
classes.

I've tested that building with Qt 4.8.6 still works.

Unfortunately qt-mustache fails about half of the tests :( To see
the failures, uncomment the QEXPECT_FAIL(..) calls in the test. I
haven't looked into fixing any of these failures, but I think that
at least some of them shouldn't be too hard.

[1] https://github.com/mustache/spec