Fix treatment of standalone tags.
Lines containing ^, /, !, <, or = tags should be stripped from the
output if the tags are standalone. The least intrusive way to handle
this I could find was to simply "expand" the endpoints of such tags
to fill the line - all existing logic will continue to just work.

This fixes 29 spec test failures.